refactoring get case interactor, writing tests
