+++ shared/src/business/useCases/getCaseInteractor.js
@@ -1,4 +1,7 @@
 const {
+  getDocumentContentsForDocuments,
+} = require('../useCaseHelper/case/getDocumentContentsForDocuments');
+const {
   isAuthorized,
   ROLE_PERMISSIONS,
 } = require('../../authorization/authorizationClientService');
@@ -6,34 +9,6 @@ const { Case, isAssociatedUser } = require('../entities/cases/Case');
 const { caseSealedFormatter } = require('../utilities/caseFilter');
 const { NotFoundError, UnauthorizedError } = require('../../errors/errors');
 const { PublicCase } = require('../entities/cases/PublicCase');
-
-export const getDocumentContentsForDocuments = async ({
-  applicationContext,
-  documents,
-}) => {
-  for (const document of documents) {
-    if (document.documentContentsId) {
-      const documentContentsFile = await applicationContext
-        .getPersistenceGateway()
-        .getDocument({
-          applicationContext,
-          documentId: document.documentContentsId,
-          protocol: 'S3',
-          useTempBucket: false,
-        });
-
-      const documentContentsData = JSON.parse(documentContentsFile.toString());
-      document.documentContents = documentContentsData.documentContents;
-      document.draftState = {
-        ...document.draftState,
-        documentContents: documentContentsData.documentContents,
-        richText: documentContentsData.richText,
-      };
-    }
-  }
-
-  return documents;
-};
 
 /**
  * getCaseInteractor
