+++ web-client/integration-tests-public/unauthedUserSearchesForCase.test.js
@@ -21,6 +21,7 @@ import { docketClerkServesOrder } from '../integration-tests/journey/docketClerk
 import { unauthedUserNavigatesToPublicSite } from './journey/unauthedUserNavigatesToPublicSite';
 import { unauthedUserSearchesByDocketNumber } from './journey/unauthedUserSearchesByDocketNumber';
 import { unauthedUserSearchesByMeta } from './journey/unauthedUserSearchesByMeta';
+import { unauthedUserSearchesForSealedCasesByDocketNumber } from './journey/unauthedUserSearchesForSealedCasesByDocketNumber';
 import { unauthedUserViewsCaseDetail } from './journey/unauthedUserViewsCaseDetail';
 import { unauthedUserViewsPrintableDocketRecord } from './journey/unauthedUserViewsPrintableDocketRecord';
 
@@ -84,6 +85,7 @@ describe('Unauthed user searches for a case and views a case detail page', () =>
   unauthedUserNavigatesToPublicSite(test);
   unauthedUserSearchesByMeta(test);
   unauthedUserSearchesByDocketNumber(test, testClient);
+  unauthedUserSearchesForSealedCasesByDocketNumber(test);
   unauthedUserViewsCaseDetail(test);
   unauthedUserViewsPrintableDocketRecord(test);
 });
